requirements.txt
langchain
langchain_groq
python-dotenv
ipykernel
fastapi
uvicorn
langserve
langchain_community